requirements.txt
python-telegram-bot==20.8
deep-translator==1.11.4
